// build/type.ts
// Generated by https://quicktype.io

export interface Root {
  title: string;
  description: string;
  data: {
    total: number;
    data: Weekly[];
  },
}

export interface Weekly {
  id:      number;
  title:   string;
  createdAt:    string;
  editor:  string;
  content: Content;
}

export interface Content {
  top:     Articles;
  article: Articles;
  library: Libraries;
}

export interface Articles {
  title: string;
  data:  Article[];
}

export interface Article {
  name:    string;
  url:     string;
  author:  Author;
  summary: string;
  tags:    Tag[];
}

export type Tag = string;

export interface Author {
  platform: string;
  column:   string;
  person:   string;
}

export interface Libraries {
  title: string;
  data:  Library[];
}

export interface Library {
  name:        string;
  url:         string;
  description: string;
  language:    string;
  tags:        string[];
}